New host file refresh code
Switch from using the DownloadManager to writing the
host files ourselves, and use AtomicFile() to do the
writing. This ensures that even if an update failed,
we can still revert to the previous state.

There is a minor race condition: If connectivity changes
during a refresh, some data from that run might be ignored,
as AtomicFile moves the backup file to the original name
during read. That's not a huge issue though, it's unlikely
to happen.

This code is tested for most cases. It's mostly the UI code
that is untested, so I hope there are no issues.

I hope that this fixes #43 - it was never entirely clear
why the host files would just vanish on their own, but now
that we write them ourselves, I'd hope they stay.

Also drop WRITE_EXTERNAL_STORAGE, this is no longer needed:
The getExternalFilesDir() does not need that permission since
API level 19 (and our minimum is 21). It was just there because
of a bug in Android's DownloadManager before 6.0 or so.