Fix support for external storage
This makes external storage work. You can specify locations in the
form of file:/absolute/path or you can specify a relative path, which
will be taken relative to the getExternalStorageDirectory() directory
(AKA /sdcard).

Permissions are only requested when there actually is a file: entry
configured, and not ignored. After a permission request, you have
to click the start button again, it does not automatically start
once you have given permission (because I'm lazy, and that's a pro
feature).

This is mostly a stop-gap: I want to access host files via content
providers instead, that also paves the way for storing them on
services like Google Drive.